- These changes remove the dependency of the blockchain backend from the diff_adj_manager and changes the calculation of the median timestamp and target difficulty to rather need the set of headers that should be used to perform the calculation. - The diff_adj_manager and diff_adj_storage was replaced with the stateless get_median_timestamp and get_target_difficulty functions that process the provided headers to calculate the median timestamp and target difficulty without requiring access to the blockchain db or backend. - The consensus manager previously exposed these methods and will query the last set of 11 headers to calculate the median timestamp and the full set of headers for the target difficulty calculation. - The median timestamp calculation will be much more efficient, but the target difficulty calculation can become more expensive as the blockchain becomes large. An improvement would be to keep the target difficulty at the horizon height as the initial difficulty for the LinearWeightedMovingAverage, then only the header set from horizon_height+1 to height needs to be requested and processed. - Added a fetch_headers command to the blockchain db. - The diff_adj_manager tests were split, all the median timestamp tests were moved to median_timestamp and all the target difficulty tests were moved to target_difficulty. - The unused test_full_sync_on_reorg test was removed.
