+++ base_layer/core/src/proof_of_work/mod.rs
@@ -21,13 +21,14 @@
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 mod blake_pow;
-mod diff_adj_manager;
 mod difficulty;
 mod error;
+mod median_timestamp;
 #[allow(clippy::enum_variant_names)]
 mod monero_rx;
 #[allow(clippy::module_inception)]
 mod proof_of_work;
+mod target_difficulty;
 
 #[cfg(test)]
 pub use blake_pow::test as blake_test;
@@ -35,8 +36,9 @@ pub use blake_pow::test as blake_test;
 pub mod lwma_diff;
 
 pub use blake_pow::{blake_difficulty, blake_difficulty_with_hash};
-pub use diff_adj_manager::{DiffAdjManager, DiffAdjManagerError};
 pub use difficulty::{Difficulty, DifficultyAdjustment};
 pub use error::{DifficultyAdjustmentError, PowError};
+pub use median_timestamp::get_median_timestamp;
 pub use monero_rx::monero_difficulty;
 pub use proof_of_work::{PowAlgorithm, ProofOfWork};
+pub use target_difficulty::get_target_difficulty;
